# services/jurisdiction-rules-svc/openapi.yaml
openapi: "3.1.0"
info:
  title: Jurisdiction Rules Service
  version: "1.1.0"
  description: |
    Authoritative registry of jurisdiction identities and their effective-dated
    applicability rules (03-microservices.md §8.2).

    **Governance invariants:**
    - `rule_payload` holds APPLICABILITY METADATA ONLY — never computation values.
      Thresholds, rates and bands belong to the Tax and Payroll services (OQ-1, Model B).
    - Every type/status/domain field is a VARCHAR tag stored as data. New jurisdiction
      types, rule domains and status values are added by data migration, never by a
      code change (OQ-3).
    - Nothing is deleted. Jurisdictions are deactivated with `active_flag` + `effective_to`;
      rules are end-dated and moved to SUPERSEDED or RETIRED.
    - Historical actions must always be explainable against the rule set active at the
      time of execution, so point-in-time reads keep working after a jurisdiction is
      deactivated and after a rule is superseded.
    - `legal_drift_state` transitions are append-only. The column carries the current
      value; `/v1/rules/{id}/drift-events` carries the history (OQ-4).
    - Jurisdiction and rule data are platform-wide reference data. There is no
      `tenant_id` and no row-level security on these tables.

    **Read vs runtime resolution.** `/rules` is the raw per-jurisdiction, effective-dated
    view — the audit and replay surface, one jurisdiction only. `/rule-pack` is the
...

...

...
servers:
...
paths:
...
  /v1/packs/{pack_ref}/versions/{version}/certification:
    get:
      tags: [Packs]
      summary: The certification record with a live signature verdict
      operationId: getCertification
      parameters:
        - { name: pack_ref, in: path, required: true, schema: { type: string } }
        - { name: version, in: path, required: true, schema: { type: string } }
      responses: { "200": { description: "{certification, verification}." }, "404": { $ref: '#/components/responses/NotFound' } }

  # ── ZS-JUR-001 Wave 2: runtime rule resolution ──────────────────────────────
...
components:
  securitySchemes:
    principalHeader:
      type: apiKey
      in: header
      name: X-Principal-Id
      description: |
        The acting principal, set by the gateway's ForwardAuth middleware after it
        verifies the caller's identity envelope against identity-svc's JWKS. Callers
        do not set this themselves — a request that reaches the service without it is
        rejected with `401 missing_principal` rather than attributed to the platform.
...
  responses:
...
    NotFound:
      description: |
        `jurisdiction_not_found` / `rule_not_found`. Also returned for a malformed id,
        which cannot name an existing row.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [type, title, status, detail, instance]
      description: |
        RFC 9457 Problem Details format. Clients can switch on `type` or `title`.
        The `extension` fields provide machine-readable codes.
      properties:
        type:
          type: string
          format: uri
          description: URI identifying the problem type.
        title:
          type: string
          description: Short, human-readable summary of the problem type.
        status:
          type: integer
          format: int32
          description: HTTP status code generated by the origin server.
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
        instance:
          type: string
          format: uri
          description: URI reference identifying the specific occurrence.
        code:
          type: string
          description: Stable machine-readable code for client logic (e.g., "missing_field").
        field:
          type: string
          description: The offending field, when code is "missing_field".
        jurisdiction_id:
          type: string
          format: uuid
          description: The jurisdiction ID, when relevant.
...